Fixed endpoints, added missing filters and entities

+++ src/main/kotlin/com/paysera/lib/accounts/entities/preciousMetals/BullionOptionWeight.kt
@@ -2,5 +2,5 @@ package com.paysera.lib.accounts.entities.preciousMetals
 
 data class BullionOptionWeight(
     val measurementUnit: String,
-    val amount: Float
+    val amount: Double
 )
